requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-WTF>=1.2.1
python-dotenv>=1.0.1
reportlab>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
ipykernel>=6.25.0
Pillow>=10.0.0
Authlib>=1.2.1
requests>=2.31.0